Update logic related to the representation of a window
Previously this package would only show a sublime window in the list if it contained a folder. These changes make it so it will use the sublime-project name if there is a .sublime-project file in the project.

In addition, if there is a window containing no folders or sublime projects it will use the path or file contents (if not saved) of the active file in that window.

Fixes #4
Fixes #8