Add experimental switching flag for Mac
This change allows switching by using the Window menu and selecting the other
window with AppleScript. I have not tested it extensively, and I don’t believe
it is actually any faster than the current behavior, but I am throwing it here
in this branch just so I have it accessible in case I want to come back to it
in the future.